♻️  zv: Refactor `as_value` module
